# back-end/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Noroff EP e-commerce",
    "description": "APIs and a database to showcase their products for customers to purchase. Admin front-end, which uses the back-end system to manage the data."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/orders/": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Fetch the logged-in user",
        "description": "A list of the logged-in user",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserOrder"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/orders/all-orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Fetch all orders",
        "description": "A list of orders made by all users. Requires Admin authentication",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AllOrders"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/orders/{id}/": {
      "put": {
        "tags": [
          "Orders"
        ],
        "summary": "Update status of an order",
        "description": "Update the status of an order, only options available are \\'In Progress\\', \\'Ordered\\', \\'Completed\\'. Requires Admin authentication",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the order to update"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Update status to an order",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UpdateOrder"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UserOrder": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Viewing all orders from user:"
            },
            "orders": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "OrderNumber": {
                    "type": "string",
                    "example": "j8a72g31"
                  },
                  "Status": {
                    "type": "string",
                    "example": "In Progress"
                  },
                  "MembershipStatus": {
                    "type": "string",
                    "example": "Bronze"
                  },
                  "TotalPrize": {
                    "type": "string",
                    "example": "650.00"
                  },
                  "createdAt": {
                    "type": "string",
                    "example": "Purchase date"
                  },
                  "updatedAt": {
                    "type": "string",
                    "example": "Last status update"
                  },
                  "UserId": {
                    "type": "number",
                    "example": 2
                  },
                  "OrderItems": {
                    "type": "array",
                    "example": [
                      "List of products ordered"
                    ],
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "id",
                  "OrderNumber",
                  "Status",
                  "MembershipStatus",
                  "TotalPrize",
                  "createdAt",
                  "updatedAt",
                  "UserId",
                  "OrderItems"
                ]
              }
            }
          },
          "required": [
            "result",
            "orders"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
    "AllOrders": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Viewing all orders:"
            },
            "orders": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "OrderNumber": {
                    "type": "string",
                    "example": "j8a72g31"
                  },
                  "Status": {
                    "type": "string",
                    "example": "In Progress"
                  },
                  "MembershipStatus": {
                    "type": "string",
                    "example": "Bronze"
                  },
                  "TotalPrize": {
                    "type": "string",
                    "example": "650.00"
                  },
                  "createdAt": {
                    "type": "string",
                    "example": "Purchase date"
                  },
                  "updatedAt": {
                    "type": "string",
                    "example": "Last status update"
                  },
                  "UserId": {
                    "type": "number",
                    "example": 2
                  },
                  "OrderItems": {
                    "type": "array",
                    "example": [
                      "List of products ordered"
                    ],
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "id",
                  "OrderNumber",
                  "Status",
                  "MembershipStatus",
                  "TotalPrize",
                  "createdAt",
                  "updatedAt",
                  "UserId",
                  "OrderItems"
                ]
              }
            }
          },
          "required": [
            "result",
            "orders"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
    "OrderStatus": {
      "type": "object",
      "properties": {
        "status": {
          "type": "array",
          "example": [
            "In Progress",
            "Ordered",
            "Completed"
          ],
          "items": {
            "type": "string"
          }
        }
      },
      "required": [
        "status"
      ]
    },
    "UpdateOrder": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Order has been updated"
            },
            "orders": {
              "type": "array",
              "example": [
                "In Progress",
                "Ordered",
                "Completed"
              ],
              "items": {
                "type": "string"
              }
            }
          },
          "required": [
            "result",
            "orders"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
...
  }
}